refactor(ui): route git's confirms through the app-wide bridge (#291)
git-confirm.ts binds the git-confirm-dialog testid the e2e specs assert, so
the four call sites keep their behavior while the store and the outlet move
out of features/git. AppShell mounts ConfirmDialogHost in place of
GitConfirmDialog.
